error in return type is fixed

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+
 /**
  * print_dlistint - print each element in the doubly linked list
  * @h: head of the linked list
@@ -8,12 +9,13 @@
 size_t print_dlistint(const dlistint_t *h)
 {
 	const dlistint_t *current = h;
-	int elements = 0;
+	size_t elements = 0;
 	while (current != NULL)
 	{
 		printf("%d \n", current->n);
 		elements++;
 		current = current->next;
 	}
+
 	return elements;
 }
